Lots of things: Added initial AudioServer stuff Added RefCountedResource stuff Added component ID to Assign function mapping

+++ BasicGameEngine/ComponentID.h
@@ -1,15 +1,19 @@
 #pragma once
-
+#include <vector>
+#include <functional>
+#include <unordered_map>
+#include <string>
+#include <utility>
 /// <summary>
 /// ComponentID stores the GetID function that generates
 /// the component IDs at runtime.
 /// </summary>
 
-extern int s_componentCounter;
+extern size_t s_componentCounter;
 
 template <class T>
-int GetID()
+size_t GetID()
 {
-	static int s_componentID = s_componentCounter++;
+	static size_t s_componentID = s_componentCounter++;
 	return s_componentID;
 }
